renamed home route to front page
