+++ src/app/routes/frontPage/getInitialData/index.js
