+++ src/app/routes/frontPage/getInitialData/index.test.js
